Update API description: mention limits on returned values
While trying to export my submission metadata, I was perplexed that the request seemed to be succeeding but returning a very limited selection.
This was not covered under the rate-limiting section and it did not occur to me to go digging through the full API docs, especially since the writing of the page implies no such limits (eg. "*All* stories matching foo" / "*All* comments matching bar").
I think this is an issue for almost everyone doing nontrivial work with the API and so worth mentioning up front.